Move build info upstream
