+++ GPTifier/include/commands.hpp
@@ -1,4 +1,3 @@
 #pragma once
 
 void print_help_messages();
-void print_build_information();
